+++ client/pages/DashboardContainer.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
-import PostWorkoutContainer from "./PostWorkoutContainer";
-import Feed from "./Feed";
-import Header from "./Header";
+import PostWorkoutContainer from "../components/PostWorkoutContainer";
+import Feed from "../components/Feed";
+import Header from "../components/Header";
 import { useNavigate } from "react-router-dom";
 import Cookies from "js-cookie";
